[feature #118079689] Move ApipagesController file to Api folder
